+++ CHANGELOG.md
@@ -7,11 +7,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 
 ## 0.1.1
 
+- Use `/` to filter, `esc` to cancel (instead of `ctrl+f`)
 - Use moodle colors
 - Help bar padding
-- Add `just run` command
 - Show correct version
 - Show status bar
+- Change loading spinner
+- Internal structure improved
 
 ## 0.1.0